Utilizing the newly added traceability improvements
